refactor: 비밀 댓글 체크박스 컴포넌트 폴더 구조 변경
