+++ components/detail/secretCheckbox/index.tsx
